[air][data] move train_test_split to ray.data
