+++ python/ray/air/__init__.py
@@ -8,7 +8,6 @@ from ray.air.config import (
 )
 from ray.air.data_batch_type import DataBatchType
 from ray.air.result import Result
-from ray.air.util.datasets import train_test_split
 
 __all__ = [
     "Checkpoint",
@@ -19,5 +18,4 @@ __all__ = [
     "DatasetConfig",
     "FailureConfig",
     "CheckpointConfig",
-    "train_test_split",
 ]
