+++ python/ray/air/examples/pytorch/torch_linear_dataset_example.py
@@ -8,7 +8,7 @@ import torch.nn as nn
 
 import ray
 import ray.train as train
-from ray.air import session, train_test_split
+from ray.data import session, train_test_split
 from ray.air.result import Result
 from ray.data import Dataset
 from ray.train.batch_predictor import BatchPredictor
