+++ python/ray/air/tests/test_dataset_utils.py
@@ -3,7 +3,7 @@ import pytest
 import ray
 import ray.data
 
-from ray.air import train_test_split
+from ray.data import train_test_split
 
 
 @pytest.fixture
