+++ python/ray/data/preprocessors/__init__.py
@@ -21,6 +21,8 @@ from ray.data.preprocessors.tokenizer import Tokenizer
 from ray.data.preprocessors.transformer import PowerTransformer
 from ray.data.preprocessors.vectorizer import CountVectorizer, HashingVectorizer
 
+from ray.data.preprocessors.utils import train_test_split
+
 __all__ = [
     "BatchMapper",
     "Categorizer",
@@ -41,4 +43,5 @@ __all__ = [
     "StandardScaler",
     "Concatenator",
     "Tokenizer",
+    "train_test_split",
 ]
